figma-transformer: extract Kiwi decode policy
